Removed fs-extra as a dependency
