+++ packages/fs/package.json
@@ -20,11 +20,9 @@
   "author": "Willem Wyndham",
   "license": "MIT",
   "dependencies": {
-    "fs-extra": "^7.0.1",
     "glob": "^7.1.3"
   },
   "devDependencies": {
-    "@types/fs-extra": "^5.0.5",
     "@types/glob": "^7.1.1"
   }
 }
