+++ packages/fs/src/tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "extends": "../../../tsconfig.ts.json"
+}
